refactor: renamed amazon_fmeval.py to eval_algo_mapping.py
